license_portal/requirements.txt
Django==3.2
psycopg2
pytest-django
